/trunk: cisstStereoVision - PNG deserialization issue solved by forcing libPNG to read out the PNG header from the stream
git-svn-id: https://svn.lcsr.jhu.edu/cisst/trunk@2505 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ libs/include/cisstStereoVision/svlSampleImageCustom.h
@@ -263,7 +263,7 @@ public:
         std::string codec;
         int compression;
         GetEncoder(codec, compression);
-        cmnSerializeRaw(outputStream, GetType());
+        cmnSerializeRaw(outputStream, static_cast<int>(GetType()));
         cmnSerializeRaw(outputStream, GetTimestamp());
         cmnSerializeRaw(outputStream, codec);
